reafacto : tamagui config
